test: source maps
